token: Add section for SPL Token program
#### Problem

The SPL Token program is one of the most called on mainnet-beta, but it
contains many inefficiencies, including its entrypoint and serde model.
It's also a great base for optimizations.

#### Summary of changes

Add SPL Token along with the `assert_instruction_count` tests for
initialize mint, initialize account, mint to, transfer, burn, and close
account.

From here, we can optimize all we want!