token: Add Zig implementation
#### Problem

There's only a Rust implementation of SPL Token in Rosetta, but we can
write a token program in many other languages.

#### Summary of changes

Add a simple SPL Token clone in Zig. It does not have multisig support
just yet, but it has most of the functionality needed.

CU usage is much lower than the Rust version. We might be able to
improve it more if we can make pubkey comparisons cheaper. Each one
seems to use ~30 CUs currently.